requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.88.0
h11==0.14.0
idna==3.10
pydantic==1.10.21
PyJWT==2.10.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.22.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
